services/forensics_consumer/requirements.txt
aiokafka==0.8.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
psycopg2-binary==2.9.9
numpy==1.26.2
shap==0.43.0
prometheus-client==0.19.0
python-multipart==0.0.6
